Split data/ out of model/ — data layer and architecture are distinct concerns
The data layer (EveryQueryPytorchDataset, EveryQueryBatch, QueryData) evolves on a different schedule than the model architecture. Schema work in #80 will reshape the batch contract without touching the nn.Module. Separate submodules means a schema PR diffs only data/ and two call-site edges.
