+++ src/every_query/train/configs/_demo_train.yaml
@@ -21,7 +21,7 @@ query:
 
 datamodule:
   _target_: meds_torchdata.extensions.lightning_datamodule.Datamodule
-  data_class: every_query.model.dataset.EveryQueryPytorchDataset
+  data_class: every_query.data.dataset.EveryQueryPytorchDataset
   config:
     _target_: meds_torchdata.MEDSTorchDataConfig
     tensorized_cohort_dir: ???
